pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-dance>=7.1.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.88.0",
    "psycopg2-binary>=2.9.10",
    "python-telegram-bot>=22.1",
    "replit>=4.1.2",
    "requests>=2.32.4",
    "stripe>=12.2.0",
    "telegram>=0.0.1",
    "flask-login>=0.6.3",
    "oauthlib>=3.3.0",
    "pyjwt>=2.10.1",
]
